+++ app/api-v1/validators/deleteProjectResponseValidator.js
@@ -1,7 +1,7 @@
 const OpenAPIResponseValidator = require('openapi-response-validator').default
 
-const apiDocResponses = require('../../api-doc-responses')
-const apiDoc = require('../../api-doc')
+const apiDocResponses = require('../api-doc-responses')
+const apiDoc = require('../api-doc')
 
 const DELETE_PROJECT_RESPONSES = {
   201: {
