+++ app/api-v1/validators/getProjectByIdResponseValidator.js
@@ -1,7 +1,7 @@
 const OpenAPIResponseValidator = require('openapi-response-validator').default
 
-const apiDocResponses = require('../../api-doc-responses')
-const apiDoc = require('../../api-doc')
+const apiDocResponses = require('../api-doc-responses')
+const apiDoc = require('../api-doc')
 
 const GET_PROJECT_BY_ID_RESPONSES = {
   200: {
