+++ app/api-v1/validators/getProjectsResponseValidator.js
@@ -1,5 +1,5 @@
-const apiDocResponses = require('../../api-doc-responses')
-const apiDoc = require('../../api-doc')
+const apiDocResponses = require('../api-doc-responses')
+const apiDoc = require('../api-doc')
 
 const GET_PROJECTS_RESPONSES = {
   200: {
