+++ app/api-v1/validators/putProjectResponseValidator.js
@@ -1,7 +1,7 @@
 const OpenAPIResponseValidator = require('openapi-response-validator').default
 
-const apiDocResponses = require('../../api-doc-responses')
-const apiDoc = require('../../api-doc')
+const apiDocResponses = require('../api-doc-responses')
+const apiDoc = require('../api-doc')
 
 const PUT_PROJECT_RESPONSES = {
   200: {
